deleteOldUrls method --> UrlService
